automatically backed up by learn

+++ index.js
@@ -33,7 +33,7 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
-  return parseFloat(n);
+  return parseInt(n);
 }
 
 function preserveDecimal(n) {
